backend/requirements.txt
numpy 
pandas 
sqlalchemy 
pgvector
openai 
pypdf
psycopg2-binary
fastapi
uvicorn
langchain_community
langchain-groq
langgraph
langchain
langchain_experimental
langchain_core
langchain_postgres
langchain_text_splitters
psycopg[binary,pool]
sentence_transformers
python-dotenv
pydantic
python-multipart